Added interactive map, moved proto -> base
