+++ cartoframes/maps.py
@@ -1,302 +1,164 @@
-"""
-Functions and methods for interactive and static maps
-"""
-
-import pandas as pd
-
-def create_named_map(base_url, api_key, tablename):
-    """Create a default named map for later use
-
-    :param username: CARTO username
-    :type username: string
-    :param api_key: CARTO API key
-    :type api_key: string
-    :param tablename: Table in user account to create a map from
-    :type tablename: string
-    """
-    import json
-    from time import time
-    import requests
-
-    map_name = '{table}_{time}'.format(table=tablename,
-                                       time=str(time()).replace('.', '_'))
-
-    defaults = {'named_map_name': map_name,
-                'basemap_layer': basemap_config(
-                    'https://cartodb-basemaps-{s}.global'
-                    '.ssl.fastly.net/dark_all/{z}/{x}/{y}.png'),
-                'cartocss': '#layer{ marker-width: 7; marker-fill: #00F; }',
-                'tablename': tablename,
-                'west': -45,
-                'south': -45,
-                'east': 45,
-                'north': 45,
-                'labels': ''}
-
-    filled_template = get_named_map_template() % defaults
-
-    api_endpoint = ('{base_url}api/v1/map/named'
-                    '?api_key={api_key}').format(base_url=base_url,
-                                                 api_key=api_key)
-    resp = requests.post(api_endpoint,
-                         data=filled_template,
-                         headers={'content-type': 'application/json'})
-    if resp.status_code == requests.codes.ok:
-        return json.loads(resp.text)['template_id']
-    else:
-        resp.raise_for_status()
-
-    return None
+import json
 
 
-def get_bounds(self, debug=False):
-    """Get the southwestern most point, and northeastern most point
-
-    :returns: Dictionary with the following keys:
-        * ``west``: longitude of the western-most extent
-        * ``south``: latitutde of the southern-most extent
-        * ``east``: longitude of the eastern-most extent
-        * ``north``: latitutde of the northern-most extent
-    """
-
-    query = '''
-        SELECT
-          ST_XMin(ext) as west,
-          ST_YMin(ext) as south,
-          ST_XMax(ext) as east,
-          ST_YMax(ext) as north
-        FROM (SELECT ST_Extent(the_geom) as ext
-              FROM {tablename}) as _wrap
-    '''.format(tablename=self.get_carto_tablename())
-    bounds = self.carto_sql_client.send(query)
-    if debug: print(bounds['rows'][0])
-    return bounds['rows'][0]
+def non_basemap_layers(layers):
+    return [layer for layer in layers if not layer.is_basemap]
 
 
-def _get_static_snapshot(self, cartocss, basemap, figsize=(647, 400),
-                         center=None, zoom=None, debug=False):
-    """Update a named map with a new configuration.
-
-    :param cartocss: CartoCSS to define new map style
-    :type cartocss: string
-    :param basemap: Basemap XYZ template to include in new map style
-    :type basemap: string
-    :param figsize: (Optional) Dimensions of output image (width, height) as a tuple
-    :type figsize: tuple of integers
-
-    :returns: URL to newly updated image
-    :rtype: string
-    """
-    import requests
-    try:
-        from urllib.parse import urlencode
-    except ImportError:
-        from urllib import urlencode
-
-    if isinstance(basemap, list) and len(basemap) == 2:
-        baselayer = basemap[0]
-        labels = basemap[1]
-    else:
-        baselayer = basemap
-        labels = None
-
-    map_params = {'named_map_name': self.get_carto_namedmap(),
-                  'basemap_layer': basemap_config(baselayer),
-                  'tablename': self.get_carto_tablename(),
-                  'labels': (',' + basemap_config(labels)) if labels else '',
-                  'cartocss': cartocss}
-
-    bounds = self.get_bounds()
-    args = dict(map_params, **bounds)
-    new_template = get_named_map_template() % args
-    if debug: print(new_template)
-    endpoint = ("{base_url}api/v1/map/named/"
-                "{map_name}?api_key={api_key}").format(
-                    base_url=self.get_carto_base_url(),
-                    map_name=self.get_carto_namedmap(),
-                    api_key=self.get_carto_api_key())
-
-    resp = requests.put(endpoint,
-                        data=new_template,
-                        headers={'Content-Type': 'application/json'})
-    # TODO: replace with bounding box extent instead
-    #       do this in the map creation/updating?
-    # https://carto.com/docs/carto-engine/maps-api/named-maps#arguments
-
-    if resp.status_code == requests.codes.ok:
-        mapview = {}
-        if zoom:
-            mapview['zoom'] = zoom
-        if center:
-            mapview['lon'] = center[0]
-            mapview['lat'] = center[1]
-
-        img = ("{base_url}api/v1/map/static/named/"
-               "{map_name}/{width}/{height}.png"
-               "?{params}")
-
-        return img.format(base_url=self.get_carto_base_url(),
-                          map_name=self.get_carto_namedmap(),
-                          width=figsize[0],
-                          height=figsize[1],
-                          params=urlencode(mapview))
-    else:
-        resp.raise_for_status()
-
-pd.DataFrame._get_static_snapshot = _get_static_snapshot
-pd.DataFrame.get_bounds = get_bounds
+def has_time_layer(layers):
+    return any(layer.time for layer in layers if not layer.is_basemap)
 
 
-def get_named_map_template():
-    """Return named map template. See CARTO's Maps API documentation for more information: https://carto.com/docs/carto-engine/maps-api/named-maps
+def get_map_name(layers, *, has_zoom):
+    version    = '20170406'
+    num_layers = len(non_basemap_layers(layers))
+    has_labels = len(layers) > 1 and layers[-1].is_basemap
+    has_time   = has_time_layer(layers)
 
-    :returns: named map template, with values to be filled in
-    """
+    return ('cartoframes_ver{version}'
+            '_layers{layers}'
+            '_time{has_time}'
+            '_baseid{baseid}'
+            '_labels{has_labels}'
+            '_zoom{has_zoom}').format(
+                version=version,
+                layers=num_layers,
+                has_time=('1' if has_time else '0'),
+                # [BUG] Remove this once baselayer urls can be passed in named map config
+                baseid=('1' if layers[0].source == 'dark' else '0'),
+                has_labels=('1' if has_labels else '0'),
+                has_zoom=('1' if has_zoom else '0')
+            )
 
-    template = '''{
-      "version": "0.0.1",
-      "name": "%(named_map_name)s",
-      "auth": {
-        "method": "open"
-      },
-      "placeholders": {
-      },
-      "view": {
-        "bounds": {
-          "west": %(west)s,
-          "south": %(south)s,
-          "east": %(east)s,
-          "north": %(north)s
+
+def get_map_template(layers, *, has_zoom):
+    num_layers = len(non_basemap_layers(layers))
+    has_time   = has_time_layer(layers)
+    name = get_map_name(layers, has_zoom=has_zoom)
+
+    # Add basemap layer
+    layers_field = [{
+        'type': 'http',
+        'options': {
+            # [BUG] Remove this once baselayer urls can be passed in named map config
+            'urlTemplate': layers[0].url,
+            # 'urlTemplate': '<%= basemap_url %>',
+            'subdomains': "abcd",
+        },
+    }]
+
+    # [BUG] Remove this once baselayer urls can be passed in named map config
+    placeholders = {}
+    # placeholders = {
+    #     'basemap_url': {
+    #         'type': 'sql_ident',
+    #         'default': ('https://cartodb-basemaps-{s}.global.ssl.fastly.net/'
+    #                     'dark_all/{z}/{x}/{y}.png'),
+    #     },
+    # }
+
+    for idx in range(num_layers):
+        layers_field.extend([{
+            'type': 'torque' if (has_time and idx == (num_layers - 1)) else 'mapnik',
+            'options': {
+                'cartocss_version': '2.1.1',
+                'cartocss': '<%= cartocss_{idx} %>'.format(idx=idx),
+                'sql': '<%= sql_{idx} %>'.format(idx=idx),
+                # [BUG] No [] for templating
+                # 'interactivity': '<%= interactivity_{idx} %>'.format(idx=idx),
+            }
+        }])
+        placeholders.update({
+            'cartocss_{idx}'.format(idx=idx): {
+                'type': 'sql_ident',
+                'default': ('#layer {'
+                            ' marker-fill: red;'
+                            ' marker-width: 5;'
+                            ' marker-allow-overlap: true;'
+                            ' marker-line-color: #000; }'),
+            },
+            'sql_{idx}'.format(idx=idx): {
+                'type': 'sql_ident',
+                'default': ("SELECT ST_PointFromText('POINT(0 0)', 4326) as the_geom,"
+                            " 1 as cartodb_id,"
+                            " ST_PointFromText('Point(0 0)', 3857) as the_geom_webmercator"),
+            },
+            # [BUG] No [] for templating
+            # 'interactivity_{idx}'.format(idx=idx): {
+            #     'type': 'sql_ident',
+            #     'default': '["cartodb_id"]',
+            # },
+        })
+
+    # Add labels if they're in front
+    if num_layers > 0 and layers[-1].is_basemap:
+        layers_field.extend([{
+            'type': 'http',
+            'options': {
+                # [BUG] Remove this once baselayer urls can be passed in named map config
+                'urlTemplate': layers[-1].url,
+                # 'urlTemplate': '<%= basemap_url %>',
+                'subdomains': "abcd",
+            },
+        }])
+
+    if has_zoom:
+        view = {
+            'zoom': '<%= zoom %>',
+            'center': {
+                'lng': '<%= lng %>',
+                'lat': '<%= lat %>',
+            },
         }
-      },
-      "layergroup": {
-        "version": "1.3.0",
-        "layers": [
-          %(basemap_layer)s,
-          {
-            "type": "mapnik",
-            "options": {
-              "sql": "select * from %(tablename)s",
-              "cartocss": "%(cartocss)s",
-              "cartocss_version": "2.3.0"
-            }
-          }%(labels)s
-        ]
-      }
-    }'''
-    return template
-
-def basemap_config(basemap_url):
-    template = '''
-          {
-            "type": "http",
-            "options": {
-              "urlTemplate": "%(basemap)s",
-              "subdomains": [
-                "a",
-                "b",
-                "c"
-              ]
-            }
-          }
-    '''
-    return template % {'basemap': basemap_url}
-
-
-def get_named_mapconfig(username, mapname, ):
-    """Named Maps API template for carto.js
-
-    :param username: The username of the CARTO account
-    :param mapname: The mapname a cartoframe is associated with CARTO account
-
-    :returns: mapconfig object for a named map as serialized JSON
-    """
-    map_args = {'mapname': mapname,
-                'username': username}
-
-    mapconfig = '''{
-      "user_name": "%(username)s",
-      "type": "namedmap",
-      "named_map": {
-        "name": "%(mapname)s"
-      },
-      "subdomains": [ "a", "b", "c" ]
-      }''' % map_args
-    return mapconfig
-
-
-def get_basemap(self, options):
-    """
-    :param options: This can be one of the following:
-
-    * XYZ URL for a custom basemap. See https://leaflet-extras.github.io/leaflet-providers/preview/ for examples.
-    * CARTO basemap style
-
-      - Specific description: `light_all`, `light_nolabels`, `dark_all`, or `dark_nolabels`
-      - General descrption: `light` or `dark`. Specifying one of these results in the best basemap for the map geometries.
-
-    * Dictionary with the following keys:
-
-      - `style`: (required) descrption of the map type (`light` or `dark`)
-      - `labels`: (optional) Show labels (`True`) or not (`False`). If this option is not included, the best basemap will be chosen based on what was entered for `style` and the geometry type of the basemap.
-
-    :type options: string or dict
-
-    :returns: basemap(s) URLs given the input parameters
-    :rtype: string or list
-    """
-
-    template = ('http://cartodb-basemaps-{{s}}.global.ssl.fastly.net/'
-                '{style}/{{z}}/{{x}}/{{y}}.png')
-    style_options = ('light_all', 'dark_all', 'light_nolabels',
-                     'dark_nolabels',)
-    label_options = ('dark', 'light',)
-
-    if isinstance(options, str):
-        if options[0:4] == 'http':
-            # input is already a basemap
-            return (options, None)
-        elif options in style_options:
-            # choose one of four carto types
-            return (template.format(style=options),
-                    'dark' if 'dark_' in options else 'light')
-        elif options in ('light', 'dark'):
-            if self.get_carto_geomtype() in ('point', 'line'):
-                return (template.format(style=options + '_all'),
-                        options)
-            else:
-                return [template.format(style=options + '_nolabels'),
-                        template.format(style=options + '_only_labels')], options
-        else:
-            raise ValueError("Text inputs must be an XYZ basemap format, or "
-                             "one of: {}.".format(','.join(style_options)))
-    elif (isinstance(options, dict) and
-          'style' in options and
-          options['style'] in label_options):
-        if ('labels' in options and
-                options['labels'] is True):
-            return ([template.format(style=options['style'] + '_nolabels'),
-                     template.format(style=options['style'] + '_only_labels')],
-                    options['style'])
-        elif ('labels' in options and options['labels'] is False):
-            return (template.format(style=options['style'] + '_nolabels'),
-                    options['style'])
-        else:
-            if self.get_carto_geomtype() in ('point', 'line'):
-                return (template.format(style=options['style'] + '_all'),
-                        options['style'])
-            else:
-                return [template.format(style=options['style'] + '_nolabels'),
-                        template.format(style=options['style'] + '_only_labels')], options['style']
+        placeholders.update({
+            'zoom': {
+                'type': 'number',
+                'default': 3,
+            },
+            'lng': {
+                'type': 'number',
+                'default': 0,
+            },
+            'lat': {
+                'type': 'number',
+                'default': 0,
+            },
+        })
     else:
-        if self.get_carto_geomtype() in ('point', 'line'):
-            return (template.format(style='dark_all'),
-                    'dark')
-        elif self.get_carto_geomtype() == 'polygon':
-            return ([template.format(style='dark_all'),
-                    template.format(style='dark_only_labels')],
-                    'dark')
-        else:
-            return (template.format(style='dark_all'),
-                    'dark')
-    return None
+        view = {
+            'bounds': {
+                'west' : '<%= west %>',
+                'south': '<%= south %>',
+                'east' : '<%= east %>',
+                'north': '<%= north %>',
+            },
+        }
+        placeholders.update({
+            'west': {
+                'type': 'number',
+                'default': -45,
+            },
+            'south': {
+                'type': 'number',
+                'default': -45,
+            },
+            'east': {
+                'type': 'number',
+                'default':  45,
+            },
+            'north': {
+                'type': 'number',
+                'default':  45,
+            },
+        })
+
+    return json.dumps({
+        'version': '0.0.1',
+        'name': name,
+        'placeholders': placeholders,
+        'layergroup': {
+            'version': '1.0.1',
+            'layers': layers_field,
+        },
+        'view': view,
+    })
