--- cartoframes/proto/__init__.py
@@ -1,3 +1,0 @@
-from .context import CartoContext
-from .layer import BaseMap, QueryLayer, Layer
-from .styling import BinMethod
